requirement.txt
google-generativeai
streamlit==1.22.0
altair<5